feat: move db middleware to separate file
