+++ templates/base-openapi/src/index.ts
@@ -1,35 +1,16 @@
 import { createFiberplane } from "@fiberplane/hono";
-import { neon } from "@neondatabase/serverless";
 import { eq } from "drizzle-orm";
-import { type NeonHttpDatabase, drizzle } from "drizzle-orm/neon-http";
 import { Hono } from "hono";
 import { describeRoute, openAPISpecs } from "hono-openapi";
 import { resolver } from "hono-openapi/zod";
-import { createMiddleware } from "hono/factory";
 import { HTTPException } from "hono/http-exception";
 import * as schema from "./db/schema";
 import { ZUserByIDParams, ZUserInsert, ZUserSelect } from "./dtos";
+import { dbProvider } from "./middleware/dbProvider";
 import { zodValidator } from "./middleware/validator";
 
-const initDb = createMiddleware<{
-  Bindings: {
-    DATABASE_URL: string;
-  };
-  Variables: {
-    db: NeonHttpDatabase;
-  };
-}>(async (c, next) => {
-  const client = neon(c.env.DATABASE_URL);
-  const db = drizzle(client, {
-    casing: "snake_case",
-  });
-
-  c.set("db", db);
-  await next();
-});
-
 const api = new Hono()
-  .use("*", initDb)
+  .use("*", dbProvider)
   .get(
     "/users",
     describeRoute({
