+++ templates/base/tests/setup.ts
@@ -1,5 +1,6 @@
 import { neon } from "@neondatabase/serverless";
 import { drizzle } from "drizzle-orm/neon-http";
+import { createMiddleware } from "hono/factory";
 import { beforeAll, afterAll, vi } from "vitest";
 import { createTestBranch, deleteBranch } from "./helpers";
 
@@ -9,14 +10,16 @@ const {
 } = await createTestBranch();
 
 beforeAll(async () => {
-  vi.mock("../src/db", () => {
+  vi.mock("../src/middleware/dbProvider.ts", () => {
     return {
-      getDb: () => {
+      dbProvider: createMiddleware(async (c, next) => {
         const db = drizzle(neon(testBranchUri), {
           casing: "snake_case",
         });
-        return db;
-      }
+
+        c.set('db', db);
+        await next();
+      }),
     }
   });
 });
